Main/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiopokeapi==0.1.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
cachelib==0.9.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
Flask==2.2.3
Flask-Caching==2.0.2
frozenlist==1.3.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
oauthlib==3.2.2
orjson==3.8.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
